Update dependencies and fix import in Motion component
