+++ package.json
@@ -8,10 +8,9 @@
 		"dev": "next dev",
 		"build": "next build",
 		"start": "next start",
-		"watch": "tsc --watch",
-		"update": "ncu -u && yarn install",
-		"check:format": "prettier --check .",
-		"check:types": "tsc"
+		"lint": "prettier .",
+		"types": "tsc",
+		"update": "ncu -u && yarn install"
 	},
 	"repository": "https://github.com/koushikpuppala/koushikpuppala.git",
 	"bugs": "https://github.com/koushikpuppala/koushikpuppala/issues",
@@ -37,14 +36,16 @@
 		"nodemailer": "^6.9.8",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
-		"react-icons": "^5.0.0",
+		"react-icons": "^5.0.1",
 		"react-tilt": "^1.0.2",
 		"react-vertical-timeline-component": "^3.6.0",
-		"sharp": "^0.33.1",
+		"sharp": "^0.33.2",
 		"typed.js": "^2.1.0"
 	},
 	"devDependencies": {
-		"@types/node": "^20.10.8",
+		"@commitlint/cli": "^18.4.4",
+		"@commitlint/config-conventional": "^18.4.4",
+		"@types/node": "^20.11.0",
 		"@types/nodemailer": "^6.4.14",
 		"@types/react": "^18.2.47",
 		"@types/react-dom": "^18.2.18",
@@ -54,7 +55,7 @@
 		"husky": "^8.0.3",
 		"postcss": "^8.4.33",
 		"postcss-import": "^16.0.0",
-		"prettier": "^3.1.1",
+		"prettier": "^3.2.2",
 		"prettier-plugin-tailwindcss": "^0.5.11",
 		"sass": "^1.69.7",
 		"tailwindcss": "^3.4.1",
